keygen: Don't regenerate keys if they already exist

+++ recipes/keygen/keygen_1.0.bb
@@ -3,7 +3,7 @@ inherit chumbysg-git
 DESCRIPTION = "Cryptoprocessor daemon, used in lieu of a real processor"
 LICENSE = "BSD"
 
-PR = "r6"
+PR = "r7"
 
 SRC_URI = "${CHUMBYSG_GIT_HOST}/chumby-clone/${PN}${CHUMBYSG_GIT_EXTENSION};protocol=${CHUMBYSG_GIT_PROTOCOL}"
 SRCREV = "c591f5c5efbf75aaa1eebb687f8830f0ddad2b14"
@@ -30,11 +30,22 @@ do_install() {
 pkg_postinst_${PN}() {
     if test "x$D" != "x"; then exit 1; fi  # Don't do postinst on build system
 
-    ifconfig wlan0 up
-    ifconfig wlan1 up
-    keygen `fpga_ctl n | cut -d" " -f3` 00000000000000000000000000060000
-    config_util --cmd=putblock --block=cpid < /tmp/keyfile
-    /etc/init.d/chumby-cpid restart
-    rm -f /tmp/keyfile*
+    KEYS=$(config_util --cmd=getblock --block=cpid | hexdump -C | grep '00 01 00 01' | wc -l)
+    if [ -z ${KEYS} ]
+    then
+        KEYS=0
+    fi
+    if [ ${KEYS} -ge 15 ]
+    then
+        echo "Not regenerating keys.  To force a regeneration, run the following:"
+        echo "config_util --cmd=putblock --block=cpid < /dev/zero; opkg install keygen"
+    else
+        ifconfig wlan0 up
+        ifconfig wlan1 up
+        keygen `fpga_ctl n | cut -d" " -f3` 00000000000000000000000000060000
+        config_util --cmd=putblock --block=cpid < /tmp/keyfile
+        /etc/init.d/chumby-cpid restart
+        rm -f /tmp/keyfile*
+    fi
     (while [ -e /usr/lib/opkg/lock ]; do sleep 1; done; opkg remove keygen)&
 }
